Created World class that will hold the List of all the location objects in the game world. Contains to functions to add a location to the location list. As will as retrieve a location from the location list if it exists within the list a NULL is returned if location isn't found. Added a World factory class. The factory will create a World and populate the world object with location objects by using the addlocation function of the world class. then return the world. The location instaiation code was removed from the gamesession class.
NOTES: The factory design pattern was used because creating the world object directly within the class that requires (uses) the object, in this case the gamesession class, commits the gamession class  to a particular object and makes it impossible to change the instantiation independently from (without having to change) the class.
Enables subclasses to change the way an object is created (to redefine which class to instantiate). GOF def: "Define an interface for creating an object, but let subclasses decide which class to instantiate. The factory method les a class defer instantiation it uses to subclasses.