NOTE:!!!!!! This is broken code. WIthin Gamession class . the CurrentLocation constructor is incorrect.  the line _currentTrader = CurrentLocation.TraderHere is WRONG!! Doing this push to document a bad bug that was not easily seen as debugging output gives no indication that there is a problem. I Believe the issue is that even though a trader is correctly assigned, the value is garbage collected as it goes out of scope once whatever MoveX() function is finished.  By using the corrected line CurrentTrader = CurrentLocation.TraderHere; a new Trader object is created that doesn't go out of scope. I think???
Added trader class.
Added traderfactory.
Location:
Added functions to get and set a trader in a location
WorldFactory:
added traders at locations (-1,0) , (0,1), (-1,-1)
//TODO figure out a way to link words with the coords.
Gamession:
Added CurrentTrader property variable
Added call to get the currentTrader in the currentLocation constructor.
Added function to get and set CurrentTrader which boardcast properchange.
Added function that returns if the location has a trader that will be used to control the visibility of a trade button by the UI.

MainWIndow:
Added trade button. Visability is controled by the HasTrader bool in Gamesession.