Created LivingEntity Class
Moved the shared properties of teh monster, player and trader class to this class.

Note: LivingEntity has the scope of “abstract”. That means you cannot instantiate a LivingEntity class. You can instantiate one of its children, which will have all the functions and properties from LivingEntity. But, you cannot create a LivingEntity object on its own.
Note: the LivingEntity constructor is declared as “protected”. Something that is protected can only been seen/used by the children of this class.

Trader.cs, Player.cs, Monster.cs:
removed properties that are now located in the livingEnity class.

MonsterFactory:
Changed to use GameItem in the inventory. as the inventory is now shared.

GameSession.cs:
changed values to match the property names used in the LivingEnity class.

MainWIndow.xaml:
changed values to match the property names used in the LivingEnity class.