Gamesession.cs: Moved logic that checked if monster or player was killed to event-handling. Updated calls to currentplayer to use new class functions.
Note: had to change the order of some of the RaiseMessage lines. Before, they were after applying damage to the player and monster. However, with the new event-handling functions, this could lead to a problem.

For example, if lines 282 and 283 were reversed (their previous order), the player could take damage and be killed. The event-handling code moves the player to their home, if they are killed. Because they are at a new location, we try to get the monster at that location. But there is no monster at the player’s home. So, CurrentMonster gets set to null.

This would cause an error with the RaiseMessage that is now on line 282. There is no “CurrentMonster”, so we can’t display CurrentMonster.Name.

So now, we have the RaiseMessage lines before we apply any changes to an object.

Tradescreen.cs:
Updated calls to currentplayer to use new class functions.

ItemFactory.cs:
fixed typo