GameItem.cs: added a way to identify an item. Will use an enum to define words to ints added new properties Category, Min and Max damage. added new properties to constructors and clone function.
removed weapon.cs

livingEntity.cs:
changed Weapons list to use new gameItem category enum.
NOTE::!! Previously 'is' was used to check if the item in the inventory. 'is' represents a Boolean value. When the implementation changed to using enum 'is' was no longer valid. However, intellsense through the error saying the type "Weapon" was not in type GameItem.ItemCategory. needed to change 'is' to == for the comparison.

player.cs:
removed currentweapon property

gamesession.cs:
updated various locations to use gameitem and/or new itemCategory enum. Also added property to get the currentWeapon.

bug wepon won't equip